python/geodio/core/parser/requirements.txt
antlr4-python3-runtime>=4.13.2
antlr4-tools>=0.2.1
antlr-denter>=1.3.1